backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
social-auth-app-django==5.4.0
Pillow==10.0.1
reportlab==4.0.4
qrcode==7.4.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-auth==2.29.0